refactor, dedup password requirement check
