+++ client/web/src/util/security.ts
@@ -1,6 +1,9 @@
 import {SourcegraphContext} from "../jscontext";
 
-export const minPasswordLen = (window.context.authMinPasswordLength !== undefined && window.context.authMinPasswordLength > 0) ? window.context.authMinPasswordLength : 12
+// make sure we always have a minPasswordLen
+export const minPasswordLen = (
+    window.context.authMinPasswordLength !== undefined && window.context.authMinPasswordLength > 0
+) ? window.context.authMinPasswordLength : 12
 
 export function getPasswordPolicy(): any {
     let passwordPolicyReference = window.context.authPasswordPolicy
@@ -18,18 +21,14 @@ export function validatePassword(
     password: string
 ): string | undefined {
 
+    // minPasswordLen always has a value so we do it first
+    if (password.length < minPasswordLen) {
+        return 'Password must be at least ' + minPasswordLen.toString() + ' characters.'
+    }
+
     let passwordPolicyReference = getPasswordPolicy()
 
     if (passwordPolicyReference?.enabled) {
-        if (
-            password.length < minPasswordLen
-        ) {
-            return (
-                'Password must be greater than ' +
-                minPasswordLen.toString() +
-                ' characters.'
-            )
-        }
         if (
             passwordPolicyReference?.numberOfSpecialCharacters &&
             passwordPolicyReference.numberOfSpecialCharacters > 0
@@ -50,7 +49,6 @@ export function validatePassword(
         }
 
         if (
-            passwordPolicyReference.requireAtLeastOneNumber &&
             passwordPolicyReference.requireAtLeastOneNumber
         ) {
             const validRequireAtLeastOneNumber = /\d+/
@@ -60,7 +58,6 @@ export function validatePassword(
         }
 
         if (
-            passwordPolicyReference.requireUpperandLowerCase &&
             passwordPolicyReference.requireUpperandLowerCase
         ) {
             const validUseUpperCase = new RegExp('[A-Z]+')
@@ -72,10 +69,41 @@ export function validatePassword(
         return undefined
     }
 
-    if (password.length < minPasswordLen) {
-        return 'Password must be at least ' + minPasswordLen.toString() + ' characters.'
-    }
-
     return undefined
 
+}
+
+export function getPasswordRequirements(
+    context: Pick<SourcegraphContext, 'authProviders' | 'sourcegraphDotComMode' | 'experimentalFeatures' |
+        'authPasswordPolicy'>
+): string {
+    let passwordPolicyReference = getPasswordPolicy()
+
+    let requirements: string  = 'Your password must include at least ' + minPasswordLen.toString() + ' characters'
+
+    if (passwordPolicyReference && passwordPolicyReference.enabled) {
+        console.log('Using enhanced password policy.')
+
+        if (
+            passwordPolicyReference.numberOfSpecialCharacters &&
+            passwordPolicyReference.numberOfSpecialCharacters > 0
+        ) {
+            requirements += ', ' + String(passwordPolicyReference.numberOfSpecialCharacters) + ' special characters'
+        }
+        if (
+            passwordPolicyReference.requireAtLeastOneNumber &&
+            passwordPolicyReference.requireAtLeastOneNumber
+        ) {
+            requirements += ', at least one number'
+        }
+        if (
+            passwordPolicyReference.requireUpperandLowerCase &&
+            passwordPolicyReference.requireUpperandLowerCase
+        ) {
+            requirements += ', at least one uppercase letter'
+        }
+
+    }
+
+    return requirements
 }
